More styling to mobile version of app
More styling added to the mobile version of the app to make it more
responsive to different sized screens. Currently there is a issue in
resizing the divider to be proportionate to the mobile design on the
image provided. So far I've attempted various changes to different
properties, both outside of the media query and within it. When none
of those worked, I modified the svg values. But this still does not fix
the issue. Which is that the divider when enlarged, clips outside of the
cards borders. Whereas it should be tucked within the card, yet larger
than it currently is.