More styling to mobile design version
Styling was added to the button to be more proportionate to the mobile
design image provided. The divider issue was left alone. Ones level of
knowledge simply does not understand at the moment what must be done.
I attempted to make it exact to the design, but it was mentioned that as
close to the original design as possible is okay too. So I will be
leaving the divider as is.