Modifications to CSS elements for responsiveness
Added a width of 100% to the container and then adjusted it down
in the media query in order to tweak a problem with the spacing not
being balance in the mobile version. JS file was completed to fetch and
generate new advice id and text on every dice roll and upon
opening the page.